move current query and mutation argument existence checks to graphql resolvers

+++ src/database/interfaces/graphql/mutations/ytitne.js
@@ -53,6 +53,9 @@ const CREATE__YTITNE__MUTATION__GRAPHQL__INTERFACE__DATABASE = {
 			"mutation_arguments",
 			mutation_arguments,
 		);
+		if (!mutation_arguments.attributes) {
+			throw new Error("Field 'attributes' is required");
+		}
 		return CONTROLLER__YTITNE__CONTROLLER__DATABASE.create(
 			{
 				source: _source,
@@ -120,6 +123,12 @@ const UPDATE__YTITNE__MUTATION__GRAPHQL__INTERFACE__DATABASE = {
 			"mutation_arguments",
 			mutation_arguments,
 		);
+		if (!mutation_arguments.attributes) {
+			throw new Error("Field 'attributes' is required");
+		}
+		if (!mutation_arguments.where) {
+			throw new Error("Field 'where' is required");
+		}
 		return CONTROLLER__YTITNE__CONTROLLER__DATABASE.update(
 			{
 				source: _source,
@@ -183,6 +192,9 @@ const DELETE__YTITNE__MUTATION__GRAPHQL__INTERFACE__DATABASE = {
 			"mutation_arguments",
 			mutation_arguments,
 		);
+		if (!mutation_arguments.where) {
+			throw new Error("Field 'where' is required");
+		}
 		return CONTROLLER__YTITNE__CONTROLLER__DATABASE.delete(
 			{
 				source: _source,
